tidy and restructure
