+++ ocf_datapipes/config/model.py
@@ -26,8 +26,6 @@ from pydantic import BaseModel, Field, root_validator, validator
 # nowcasting_dataset imports
 from ocf_datapipes.utils.consts import (
     AWOS_VARIABLE_NAMES,
-    DEFAULT_N_GSP_PER_EXAMPLE,
-    DEFAULT_N_PV_SYSTEMS_PER_EXAMPLE,
     NWP_VARIABLE_NAMES,
     SAT_VARIABLE_NAMES,
     NWP_PROVIDERS,
@@ -40,6 +38,9 @@ IMAGE_SIZE_PIXELS_FIELD = Field(
 )
 METERS_PER_PIXEL_FIELD = Field(2000, description="The number of meters per pixel.")
 METERS_PER_ROI = Field(128_000, description="The number of meters of region of interest.")
+
+DEFAULT_N_GSP_PER_EXAMPLE = 32
+DEFAULT_N_PV_SYSTEMS_PER_EXAMPLE = 2048
 
 logger = logging.getLogger(__name__)
 
@@ -726,114 +727,11 @@ class InputData(Base):
         )
 
 
-class OutputData(Base):
-    """Output data model"""
-
-    filepath: Union[str, Pathy] = Field(
-        Pathy("gs://solar-pv-nowcasting-data/prepared_ML_training_data/v7/"),
-        description=(
-            "Where the data is saved to.  If this is running on the cloud then should include"
-            " 'gs://' or 's3://'"
-        ),
-    )
-
-    @validator("filepath")
-    def filepath_pathy(cls, v):
-        """Make sure filepath is a Pathy object"""
-        return Pathy.fluid(v)
-
-
-class Process(Base):
-    """Pydantic model of how the data is processed"""
-
-    seed: int = Field(1234, description="Random seed, so experiments can be repeatable")
-    batch_size: int = Field(32, description="The number of examples per batch")
-    t0_datetime_frequency: pd.Timedelta = Field(
-        pd.Timedelta("5 minutes"),
-        description=(
-            "The temporal frequency at which t0 datetimes will be sampled."
-            "  Can be any string that `pandas.Timedelta()` understands."
-            "  For example, if this is set to '5 minutes', then, for each example, the t0 datetime"
-            " could be at 0, 5, ..., 55 minutes past the hour.  If there are DataSources with a"
-            " lower sample rate (e.g. half-hourly) then these lower-sample-rate DataSources will"
-            " still produce valid examples.  For example, if a half-hourly DataSource is asked for"
-            " an example with t0=12:05, history_minutes=60, forecast_minutes=60, then it will"
-            " return data at 11:30, 12:00, 12:30, and 13:00."
-        ),
-    )
-    split_method: split.SplitMethod = Field(
-        split.SplitMethod.DAY_RANDOM_TEST_DATE,
-        description=(
-            "The method used to split the t0 datetimes into train, validation and test sets."
-            " If the split method produces no t0 datetimes for any split_name, then"
-            " n_<split_name>_batches must also be set to 0."
-        ),
-    )
-
-    train_test_validation_split: List[int] = Field(
-        [10, 1, 1],
-        description=(
-            "The ratio of how the entire dataset is split. "
-            "Note different split methods interact different with these numbers"
-        ),
-    )
-
-    n_train_batches: int = Field(
-        250,
-        description=(
-            "Number of train batches.  Must be 0 if split_method produces no t0 datetimes for"
-            " the train split"
-        ),
-    )
-    n_validation_batches: int = Field(
-        0,  # Currently not using any validation batches!
-        description=(
-            "Number of validation batches.  Must be 0 if split_method produces no t0 datetimes for"
-            " the validation split"
-        ),
-    )
-    n_test_batches: int = Field(
-        10,
-        description=(
-            "Number of test batches.  Must be 0 if split_method produces no t0 datetimes for"
-            " the test split."
-        ),
-    )
-    upload_every_n_batches: int = Field(
-        16,
-        description=(
-            "How frequently to move batches from the local temporary directory to the cloud bucket."
-            "  If 0 then write batches directly to output_data.filepath, not to a temp directory."
-        ),
-    )
-
-    local_temp_path: Path = Field(
-        Path("~/temp/").expanduser(),
-        description=(
-            "This is only necessary if using a VM on a public cloud and when the finished batches"
-            " will be uploaded to a cloud bucket. This is the local temporary path on the VM."
-            "  This will be emptied."
-        ),
-    )
-
-    @validator("local_temp_path")
-    def local_temp_path_to_path_object_expanduser(cls, v):
-        """
-        Convert the local path to Path
-
-        Convert the path in string format to a `pathlib.PosixPath` object
-        and call `expanduser` on the latter.
-        """
-        return Path(v).expanduser()
-
-
 class Configuration(Base):
     """Configuration model for the dataset"""
 
     general: General = General()
     input_data: InputData = InputData()
-    output_data: OutputData = OutputData()
-    process: Process = Process()
     git: Optional[Git] = None
 
     def set_base_path(self, base_path: str):
