+++ ocf_datapipes/utils/consts/__init__.py
@@ -1,0 +1,21 @@
+"""useful global constants"""
+from .batches import (
+    BatchKey, NWPBatchKey, NumpyBatch, XarrayBatch
+)
+from .location import Location
+from .data_vars import (
+    NWP_PROVIDERS, 
+    UKV_VARIABLE_NAMES,
+    UKV_STD,
+    UKV_MEAN,
+    GFS_VARIABLE_NAMES,
+    GFS_STD,
+    GFS_MEAN,
+    NWP_VARIABLE_NAMES,
+    NWP_STD,
+    NWP_MEAN,
+    RSS_VARIABLE_NAMES,
+    RSS_STD,
+    RSS_MEAN,
+    AWOS_VARIABLE_NAMES,
+)
